+++ packages/ui/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @calcifer/ui
+# @calcifer-design/ui
 
 ## 0.1.0
 
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies [59a2bf4]
-  - @calcifer/tokens@0.1.0
+  - @calcifer-design/tokens@0.1.0